[#21] Rename: move validateNickname to apis/user
